Tweak type definitions used in exports map for TS 4.8 compatibility
Duplicate and export shared types where appropriate, but don't export from the other export file since it seems to confuse things.
