+++ entrypoints/main.d.ts
@@ -1,2 +1,12 @@
-export * from './main.cjs';
-export {default} from './main.cjs';
+import type {TestFn} from '../types/test-fn.cjs';
+
+export * from '../types/assertions.cjs';
+export * from '../types/try-fn.cjs';
+export * from '../types/test-fn.cjs';
+export * from '../types/subscribable.cjs';
+
+/** Call to declare a test, or chain to declare hooks or test modifiers */
+declare const test: TestFn;
+
+/** Call to declare a test, or chain to declare hooks or test modifiers */
+export default test;
